Add client stripe validation

+++ SkiNet/client/src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -26,6 +26,7 @@ export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
   cardCvc: any;
   cardErrors: any;
   publishableKey: string = environment.publishableKey;
+  cardHandler = this.onChange.bind(this);
 
   constructor(private basketService: BasketService, private checkoutService: CheckoutService, 
     private toastr: ToastrService, private router: Router) { }
@@ -36,18 +37,30 @@ export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
 
     this.cardNumber = elements.create('cardNumber');
     this.cardNumber.mount(this.cardNumberElement.nativeElement);
+    this.cardNumber.addEventListener('change', this.cardHandler);
     
     this.cardExpiry = elements.create('cardExpiry');
     this.cardExpiry.mount(this.cardExpiryElement.nativeElement);
+    this.cardExpiry.addEventListener('change', this.cardHandler);
 
     this.cardCvc = elements.create('cardCvc');
     this.cardCvc.mount(this.cardCvcElement.nativeElement);
+    this.cardCvc.addEventListener('change', this.cardHandler);
   }
 
   ngOnDestroy(): void {
     this.cardNumber.destroy();
     this.cardExpiry.destroy();
     this.cardCvc.destroy();
+  }
+
+  onChange({error}) {
+    if (error) {
+      this.cardErrors = error.message;
+    }
+    else {
+      this.cardErrors = null;
+    }
   }
 
   submitOrder() {
@@ -57,9 +70,24 @@ export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
     this.checkoutService.createOrder(orderToCreate).subscribe({
       next: (order: IOrder) => {
         this.toastr.success('Order created successfully');
-        this.basketService.deleteLocalBasket(basket.id);
-        const navigationExtras: NavigationExtras = {state: order};
-        this.router.navigate(['checkout/success'], navigationExtras);
+        this.stripe.confirmCardPayment(basket.clientSecret, {
+          payment_method: {
+            card: this.cardNumber,
+            billing_details: {
+              name: this.checkoutForm.get('paymentForm').get('nameOnCard').value
+            }
+          }
+        }).then((result: any) => {
+          console.log(result);
+          if (result.paymentIntent) {
+            this.basketService.deleteLocalBasket(basket.id);
+            const navigationExtras: NavigationExtras = {state: order};
+            this.router.navigate(['checkout/success'], navigationExtras);
+          }
+          else {
+            this.toastr.error(result.error.message);
+          }
+        });
       },
       error: (error) => {
         this.toastr.error(error.message);
